% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{play a dice-tossing game}
\usage{
dice(...)
}
\arguments{
\item{...}{please specify all the options you're wagging amongst.}
}
\value{
Whatever it returns, gonna do it! It is said that the true point
of tossing is that, when waiting for the coin or dice in the air you finally
know your hope. hehe
}
\description{
When you don't know what to do, this helps you to make a
final decision via a simulated coin tossing or dice tossing game.
}
\examples{
dice('Do the Laundry','Homework','Cook','Go to Bed')
}
