% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvInfo.R
\name{InvInfo}
\alias{InvInfo}
\title{read Invoice Info from elect invoice pdfs}
\usage{
InvInfo(x)
}
\arguments{
\item{x}{path to the pdfs. Can be a vector for several files. Better set
working directory first.}
}
\value{
A data frame containing main information of those files.
}
\description{
read invoice pdfs and extract main information. Current version
can take into account the file name, invoice No., date and price.
}
\examples{
setwd('D://Invoice/')
Invfiles <- list.files(pattern = 'pdf$')
Inv <- InvInfo(Invfiles)
}
