% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secured.R
\name{secured}
\alias{secured}
\title{let's communicate in secret!}
\usage{
secured(x)
}
\arguments{
\item{x}{a word to be secured}
}
\value{
secured / unsecured strings.
}
\description{
to transcode a word to it's opposite in alphabet, vowel by vowel, consonant by consonant.
}
\examples{
secured('Li Xunhuan')
}
