% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{play a dice-tossing game}
\usage{
dice(Points)
}
\arguments{
\item{Points}{a character vector specifying all the options you're
wagging amongst.}
}
\description{
When you don't know what to do, this helps you to make a
final decision via a simulated coin tossing or dice tossing game.
}
\examples{
Points <- c('Do the Laundry','Homework','Cook','Go to Bed')

}
